+++ Library/src/comms/USBLSimple.cpp
@@ -20,7 +20,7 @@
 //  Stonefish
 //
 //  Created by Patryk Cieślak on 25/02/2020.
-//  Copyright (c) 2020-2021 Patryk Cieslak. All rights reserved.
+//  Copyright (c) 2020-2025 Patryk Cieslak. All rights reserved.
 //
 
 #include "comms/USBLSimple.h"
@@ -51,10 +51,12 @@ void USBLSimple::setResolution(Scalar range, Scalar angleDeg)
 
 void USBLSimple::ProcessMessages()
 {
-    AcousticDataFrame* msg;
-    while((msg = (AcousticDataFrame*)ReadMessage()) != nullptr)
+    std::shared_ptr<AcousticDataFrame> msg;
+    while((msg = std::static_pointer_cast<AcousticDataFrame>(ReadMessage())) != nullptr)
     {
-        if(msg->data == "ACK")
+        std::string ack {"ACK"};
+        std::vector<uint8_t> ackData(ack.begin(), ack.end());
+        if(msg->data == ackData)
         {  
             //Get message data
             AcousticModem* cNode = getNode(msg->source);
@@ -112,8 +114,6 @@ void USBLSimple::ProcessMessages()
 
             newDataAvailable = true;
         }
-        
-        delete msg;
     }
 }